pyproject.toml
[tool.poetry]
name = "ricetypeimagedetection"
version = "0.1.0"
description = "ML project with aim to detect rice by its image."
authors = ["TaniaZharova2205"]
readme = "README.md"
packages = [{ include = "rice_type_image_detection" }]

[tool.poetry.dependencies]
python = "^3.10"
fire = "^0.7.0"
torch = "^2.5.0"
torchvision = "^0.22.0"
pytorch-lightning = "^2.5.0"
pandas = "^2.2.3"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
kagglehub = ">=0.3.12,<0.4.0"
datasets = ">=3.6.0,<4.0.0"
transformers = ">=4.52.4,<5.0.0"
gitpython = ">=3.1.44,<4.0.0"
onnx = ">=1.18.0,<2.0.0"
onnxruntime = ">=1.22.0,<2.0.0"
tritonclient = {extras = ["http"], version = ">=2.58.0,<3.0.0"}
pycuda = ">=2025.1.2,<2026.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
mlflow = "^2.20.3"
dvc = "^3.59.1"
hydra-core = "^1.3.2"

[tool.poetry.group.inference.dependencies]
tritonclient = "^2.58.0"

[tool.isort]
src_paths = ["rice_type_image_detection"]
profile = "black"
line_length = 79
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
target-version = "py310"
src = ["rice_type_image_detection"]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E203", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
